Extend slow pod startup how-to with `fsGroupChangePolicy`
We've seen that changing the CephFS CSI driver config to use
`fsGroupPolicy: File` causes slow pod starts for CephFS volumes with
many files, because the container runtime unconditionally updates the
group permissions and ownership for all files on each pod start by
default.

Setting `fsGroupChangePolicy: OnRootMismatch` instructs the container
runtime to only update group permissions and ownership when the root
directory's group doesn't match the fsGroup of the pod. In combination
with SELinux type `spc_t`, this fixes the slow pod startup when the CSI
driver is configured with `fsGroupPolicy: File`.